pyproject.toml
[project]
name = "rag-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
]
